Plumb through user analytics
